- Fixed: Save versioning didn't work at times.  atoi was not wide enough to hold our build time stamps. - Added the version number to the save games to use as a fallback so out of repository builds without a time stamp can still roughly determine incompatible save files. - Fixed: A few warnings.

+++ src/wl_loadsave.h
@@ -41,7 +41,7 @@ class MenuItem;
 
 namespace GameSave
 {
-	extern int SaveVersion;
+	extern long long SaveVersion;
 	extern const char* savedir;
 
 	Menu		&GetLoadMenu();
